refactor: pipe, static 이벤트 파일 분리 #189
